backend/requirements.txt
accelerate==0.33.0
aiobotocore==2.20.0
aiohappyeyeballs==2.4.6
aiohttp==3.11.12
aioitertools==0.12.0
aiosignal==1.3.2
anndata==0.11.3
annotated-types==0.7.0
anyio==4.8.0
array_api_compat==1.10.0
asttokens==3.0.0
attrs==25.1.0
beautifulsoup4==4.13.3
biopython==1.85
botocore==1.36.23
cellxgene-census==1.15.0
certifi==2025.1.31
charset-normalizer==3.4.1
click==8.1.8
comm==0.2.2
contourpy==1.3.1
cycler==0.12.1
dataclasses==0.6
datasets==2.19.2
decorator==5.1.1
dill==0.3.8
et_xmlfile==2.0.0
evaluate==0.4.2
executing==2.2.0
fastapi==0.115.8
filelock==3.17.0
fonttools==4.56.0
frozenlist==1.5.0
fsspec==2024.3.1
fuzzywuzzy==0.18.0
gget==0.29.0
h11==0.14.0
h5py==3.13.0
huggingface-hub==0.29.1
idna==3.10
ipython==8.32.0
ipywidgets==8.1.5
jedi==0.19.2
Jinja2==3.1.5
jmespath==1.0.1
joblib==1.4.2
jupyterlab_widgets==3.0.13
kiwisolver==1.4.8
legacy-api-wrap==1.4.1
llvmlite==0.44.0
lxml==5.3.1
MarkupSafe==3.0.2
matplotlib==3.10.0
matplotlib-inline==0.1.7
mpmath==1.3.0
multidict==6.1.0
multiprocess==0.70.16
mysql-connector-python==9.2.0
natsort==8.4.0
networkx==3.4.2
numba==0.61.0
numpy==1.26.4
openpyxl==3.1.5
packaging==24.2
pandas==2.2.3
parso==0.8.4
patsy==1.0.1
pexpect==4.9.0
pillow==11.1.0
platformdirs==4.3.6
prompt_toolkit==3.0.50
propcache==0.3.0
psutil==7.0.0
ptyprocess==0.7.0
pure_eval==0.2.3
pyarrow==12.0.1
pyarrow-hotfix==0.6
pydantic==2.10.6
pydantic_core==2.27.2
Pygments==2.19.1
pynndescent==0.5.13
pyparsing==3.2.1
pytdc==1.1.14
python-dateutil==2.9.0.post0
pytz==2025.1
PyYAML==6.0.2
rdkit==2023.9.6
regex==2024.11.6
requests==2.32.3
s3fs==2024.3.1
safetensors==0.5.2
scanpy==1.11.0
scikit-learn==1.2.2
scipy==1.15.2
seaborn==0.13.2
session-info2==0.1.2
six==1.17.0
sniffio==1.3.1
somacore==1.0.11
soupsieve==2.6
stack-data==0.6.3
starlette==0.45.3
statsmodels==0.14.4
sympy==1.13.1
threadpoolctl==3.5.0
tiledb==0.29.1
tiledbsoma==1.11.4
tokenizers==0.19.1
torch==2.6.0
tqdm==4.67.1
traitlets==5.14.3
transformers==4.43.4
typing_extensions==4.12.2
tzdata==2025.1
umap-learn==0.5.7
urllib3==2.3.0
uvicorn==0.34.0
wcwidth==0.2.13
widgetsnbextension==4.0.13
wrapt==1.17.2
xxhash==3.5.0
yapf==0.43.0
yarl==1.18.3
